requirements.txt
beautifulsoup4==4.9.3
boto3==1.16.40
botocore==1.19.40
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.26.0
google-auth==1.25.0
google-cloud-core==1.6.0
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.0
langdetect==1.0.8
lxml==4.6.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
nltk==3.5
numpy==1.18.0
packaging==20.9
pandas==1.1.5
Pillow==8.0.1
progress==1.5
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycld3==0.20
pycparser==2.20
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
regex==2020.11.13
requests==2.25.1
rfc3986==1.4.0
rsa==4.7
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
textblob==0.15.3
tqdm==4.55.0
urllib3==1.26.2
Werkzeug==1.0.1
XlsxWriter==1.3.7
